refactor: UserDao interface로 추출 및 UserDaoImpl 구현체 구현
